Add files via upload
Write a Java program to print the result of the following operations for x = 1 and y =2. After every operation print the result of the whole operation, x and y.
a. -x + x++ * 2
b. (x++ + --y) + --y
c. x + x
At the end, after all the operations print the values of x and y. 

 

Write a Java program to print the area and perimeter of a rectangle. 
Test Data:
Width = 5.5 Height = 8.5
Expected Output
Area is 5.6 * 8.5 = 47.60
Perimeter is 2 * (5.6 + 8.5) = 28.20 

 

Write a Java program to convert seconds to hour, minute and seconds.

 

Write a program to enter names and year of birth of 3 people. Print a message for each person showing his/her details.

 

Write a Java program to convert temperature from Fahrenheit to Celsius degree.

 

Write a Java program that reads a number in inches, converts it to meters.
Note: One inch is 0.0254 meter.

 

Γράψτε ένα πρόγραμμα που δέχεται την τιμή ενός προϊόντος και εμφανίζει σε μηνύματα τη τιμή του σε έκπτωση 10%, 20%, 30%, 40%, 50%, 60%.