Add files via upload
Exercise 1
Create a program that accepts input from a user.
The program must ask the user to enter info about his/her family.
The details that we need for each person are the following:
Name, Age, Height, Weight.
Store these details in a file in your Operating System.
The format of the file must be the following:
NAME        AGE        HEIGHT(m)    WEIGHT(kg)
Jack        12            1.85            65
Mary        24            1.76            76
The program should end only when the user enters "exit". But the process cannot be interrupted while entering details of a user. Everytime, the user inserts a person, the program should inform the user that "Person was successfully created".
Finally read the file that you created and print everything in the output
***Use formatted strings to create the format of the file(String.format)
***Create methods
 
Exercise 2

Create a .txt file that contains the following 5 URLs.
https://www.javatpoint.com/
https://docs.oracle.com/javase/8/docs/
https://www.baeldung.com/
https://www.vogella.com//tutorials/java.html
https://www.tutorialspoint.com/java_technology_tutorials.htm
Create a method for the following:
Read the file and print only the subdomain of each url.
For example the subdomain of https://www.javatpoint.com/ is javatpoint.
The subdomain of https://docs.oracle.com/javase/8/docs/ is docs.oracle
Also create a method that prints only the urls that contain the word java.