Create Private School Project
This private school supports multiple courses.
Each course can have multiple trainers, students and assignments.
For example,
the Course CB8 has full time Java and C# and part time Java and C#.
The CB8 course starts on 1/1/2019 and ends at 31/3/2019.
On this CB8 course there are 20 students enrolled and 8 trainers that teach.
During the bootcamp the students need to submit five (5) assignments and two (2) projects, one
individual and one team project.
The application that you will build needs to hold in a database multiple courses along with the
enrolled students, the trainers that teach the subjects and the assignments / projects that the
students need to submit during the course duration.
This individual project will break into to parts. On the next brief you will get detailed information
on what you need to implement.